stage toolbox naming and refactoring
