+++ ui/dashboard/src/components/stage/Toolbox/tools/Scenes/BlankScene.vue
