+++ ui/dashboard/src/components/stage/Toolbox/tools/Settings.vue
