+++ ui/dashboard/src/components/stage/Toolbox/tools/Streams/index.vue
